added loading overlay
